Move Import commands to Core
